requirements.txt
Django==1.11.18
djangorestframework==3.7.7
djangorestframework-serializer-extensions==0.6.0
